fix(marshaling): shorten marshaling to enable more banks
While mentioned in the docs as possible this is not strictly necessary
and bank institutes should be able to parse the message with these
additional control characters. Unfortunately some institutes do not
parse the messages correctly when they contain these additional
characters.

One could argue that it's a bug on institute side rather than client
side.